Flying Etiquette Survey Data

This is a modified version of the Flying Etiquette Survey data behind the story: 41 percent of flyers say it's rude to recline your seat on an airplane.

Usage

flying

Format

a dataframe 1040 obs. and 28 columns, which are:

respondent_id

respondentid

travel_frequency

how often do you travel by plane?

seat_recline

do you ever recline your seat when you fly?

height

how tall are you?

children_under_18

do you have any children under 18?

two_armrests

in a row of three seats, who should get to use the two arm rests?

middle_armrest

in a row of two seats, who should get to use the middle arm rest?

window_shade

who should have control over the window shade?

moving_to_unsold_seat

is it rude to move to an unsold seat on a plane?

talking_to_seatmate

generally speaking, is it rude to say more than a few words to the stranger sitting next to you on a plane?

getting_up_on_6_hour_flight

on a 6 hour flight from nyc to la, how many times is it acceptable to get up if you're not in an aisle seat?

obligation_to_reclined_seat

is it rude to recline your seat on a plane?

eliminate_reclining_seats

given the opportunity, would you eliminate the possibility of reclining seats on planes entirely?

switch_for_friends

is it rude to ask someone to switch seats with you in order to be closer to friends?

switch_for_family

is it rude to ask someone to switch seats with you in order to be closer to family?

wake_passenger_bathroom

is it rude to wake a passenger up if you are trying to go to the bathroom?

wake_passenger_walk

is it rude to wake a passenger up if you are trying to walk around?

baby_on_plane

in general, is it rude to bring a baby on a plane?

unruly_children

in general, is it rude to knowingly bring unruly children on a plane?

electronics_violation

have you ever used personal electronics during take off or landing in violation of a flight attendant's direction?

smoking_violation

have you ever smoked a cigarette in an airplane bathroom when it was against the rules?

gender

gender

age

age

household_income

household income

education

education

location_census_region

location (census region)

survey_type

type of the survey

Source

https://github.com/fivethirtyeight/data/tree/15f210532b2a642e85738ddefa7a2945d47e2585/flying-etiquette-survey

https://fivethirtyeight.com/features/airplane-etiquette-recline-seat/

Examples

data(flying)